moved sound card drivers
